Move generateString to common file
